+++ apps/backoffice-v2/src/common/components/molecules/ProcessTracker/hooks/useProcessTracker/process-tracker-adapters/collection-flow/components/CollectionFlowStepItem/hooks/useRequestStepFromClient/set-step-state-in-context-to-revision.ts
@@ -1,0 +1,22 @@
+import { TWorkflowById } from '@/domains/workflows/fetchers';
+import {
+  CollectionFlowStepStatesEnum,
+  TCollectionFlowStep,
+  updateCollectionFlowStep,
+} from '@ballerine/common';
+
+export const updateStepStateAndReasonInContext = (
+  context: TWorkflowById['context'],
+  step: TCollectionFlowStep,
+  state: keyof typeof CollectionFlowStepStatesEnum,
+  reason: string | undefined,
+) => {
+  const contextClone = structuredClone(context);
+
+  updateCollectionFlowStep(contextClone, step.stepName, {
+    state,
+    reason,
+  });
+
+  return contextClone;
+};