+++ apps/backoffice-v2/src/domains/workflow-definitions/fetchers.ts
@@ -27,6 +27,7 @@ export const WorkflowDefinitionConfigSchema = z
     isManualCreation: z.boolean().default(false),
     isAssociatedCompanyKybEnabled: z.boolean().default(false),
     isCaseOverviewEnabled: z.boolean().default(false),
+    isCollectionFlowPageRevisionEnabled: z.boolean().default(false),
     isCaseRiskOverviewEnabled: z.boolean().default(false),
     isDocumentTrackerEnabled: z.boolean().default(false),
     theme: WorkflowDefinitionConfigThemeSchema.default({