+++ packages/ui/src/components/organisms/Form/DynamicForm/fields/FieldList/FieldList.tsx
@@ -59,15 +59,16 @@ export const FieldList: TDynamicFormField<IFieldListParams> = props => {
               <span className="text-sm font-bold">
                 {itemIndexLabel.replace('{INDEX}', (index + 1).toString())}
               </span>
-              <span
-                role="button"
+              <button
                 tabIndex={0}
-                className="cursor-pointer text-sm"
+                disabled={disabled}
+                aria-disabled={disabled}
+                className="disabled:opacity-50 text-sm font-bold"
                 onClick={() => removeItem(index)}
                 data-testid={`${fieldId}-fieldlist-item-remove-${index}`}
               >
                 {removeButtonLabel}
-              </span>
+              </button>
             </div>
             <StackProvider stack={[...(stack || []), index]}>
               <Renderer